// src/solicitud/dtos/create-solicitud-dto.ts
export class CreateSolicitudDto {
  nombre: string; // 'Sara'
  cargo: string; // 'formador'
  promocion: string; // 'p7'
  email: string;
  tipo: string; // 'masterclass'
  nombreActividad: string;
  //'Taller Testing APIs NestJS';
  start: Date;
  end: Date;
  dia: string;
  horaInicio: string;
  horaFin: string;
}